Move exporting related methods to BackendExporting
